Add files via upload
A distributed program called the Music Radio System enables publishers to upload music to a central server and users to download songs from the server. To enable remote communication between the clients and the server, Java Remote Method Invocation (RMI) technology is used in its implementation. Only authenticated users will be able to access the music collection thanks to the system's secure login support for both publishers and consumers.